libstand: cd9660 comparison between signed and unsigned integer expressions
/code/illumos-gate/usr/src/boot/lib/libstand/cd9660.c: In function 'buf_read_file':
/code/illumos-gate/usr/src/boot/lib/libstand/cd9660.c:457:14: error: comparison between signed and unsigned integer expressions [-Werror=sign-compare]
  if (*size_p > fp->f_size - fp->f_off)
              ^
cc1: all warnings being treated as errors

+++ usr/src/boot/lib/libstand/cd9660.c
@@ -486,7 +486,7 @@ buf_read_file(struct open_file *f, char **buf_p, size_t *size_p)
 	*buf_p = fp->f_buf + blkoff;
 	*size_p = ISO_DEFAULT_BLOCK_SIZE - blkoff;
 
-	if (*size_p > fp->f_size - fp->f_off)
+	if (*size_p > (size_t)(fp->f_size - fp->f_off))
 		*size_p = fp->f_size - fp->f_off;
 	return (rc);
 }
